+++ CHANGELOG.md
@@ -19,6 +19,7 @@ y este proyecto sigue [Versionado Semántico](https://semver.org/lang/es/).
 - Repositorio [Category](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/repository/CategoryRepository.kt) para la entidad category
 - Repositorio [Product](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/repository/ProductRepository.kt) para la entidad product
 - Repositorio [Purchase](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/repository/PurchaseRepository.kt) para la entidad Purchase
+- Repositorio [PurchaseDetail](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/repository/PurchaseDetailRepository.kt) para la entidad PurchaseDetail
 
 ## [0.0.1] - 2026-07-03